Move polo-specific files to top level, make this a host branch
This is in preparation to switching to dotdotdot, which supports
multiple repositories to be mixed in to each other. The idea here will
be that master will be all the common files and then I'll start using
host-specific branches again, but now all of those branches will _just_
have the files which are to be overridden.

This is different than my previous approach since the host-specific
branches no longer need to have _all_ dotfiles within them since
dotdotdot will take care of mixing in different repos/branches as
needed.